refactor(experience): reorg hooks location reorg hooks location
